omx-plugins: h/w accelerated thumbnail for H.264
Added h/w accelerated support for H.264 thumbnails.
This required adding "OMX.Intel.index.ThumbnailMode"
extension in stagefright.

Add stagefright logic to re-order components to use h/w
accelerated for thumbnails.  Currently playback is performed
by s/w OMX codec.

To be able to use the h/w accelerated plug-in the rest of the
stack has to be enabled.  If not enable the code added won't
show regression in functionality.

This patch must be filtered out when submitting this branch
upstream

Change-Id: Id9f0e15fcbcd3c19ebb590fde45983d96e9717ef
Signed-off-by: Daniel Charles <daniel.charles@intel.com>
Signed-off-by: Daniel Leung <daniel.leung@intel.com>