MUST_REVERT: Use h/w decoder for h.264 first for video playback.
This is to allow h.264 main and high profiles playback. Also
baseline profile will use the h/w decoder.

Considered as a temporary solution to allow video playback.  This
temporary solution maps (copies) decoded raw yuv nv12 frames to
user space to push them to the Surface Flinger for rendering. A
color conversion nv12 to rgb585 is performed on cpu before pushing.

This patch is not meant to go upstream.

For: AXIA-328
Depends-Change-Id: Ic75806a996f1fee3a46a02ddbd2f3e4e619e48c9
Depends-Change-Id: I77f025f9aaabfe8db37b960c04b9e7ee73670e0c
Signed-off-by: Daniel Charles <daniel.charles@intel.com>

Change-Id: I1542fb1491f8d789c714d225b554da0a39f06d5e
Signed-off-by: Daniel Charles <daniel.charles@intel.com>